+++ kitty_tests/layout.py
@@ -31,8 +31,8 @@ class Window:
     def effective_padding(self, edge):
         return 1
 
-    def effective_margin(self, edge, is_single_window=False):
-        return 0 if is_single_window else 1
+    def effective_margin(self, edge):
+        return 1
 
     def set_visible_in_layout(self, val):
         self.is_visible_in_layout = bool(val)