+++ docs/changelog.rst
@@ -52,6 +52,8 @@ Detailed list of changes
 
 - A new :doc:`escape code <pointer-shapes>` that can be used by programs running in the terminal to change the shape of the mouse pointer (:iss:`6711`)
 
+- A new option :opt:`single_window_padding_width` to use a different padding when only a single window is visible (:iss:`6734`)
+
 - A new mouse action ``mouse_selection word_and_line_from_point`` to select the current word under the mouse cursor and extend to end of line (:pull:`6663`)
 
 - Allow using the full range of standard mouse cursor shapes when customizing the mouse cursor